gamereviewer/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.6
djangorestframework==3.14.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.2.0
mysqlclient==2.2.0
protobuf==4.21.12
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
djangorestframework-simplejwt==5.3.0